requirements.txt
flask>=2.2.0
gunicorn
werkzeug
pikepdf
pillow
python-docx
reportlab
moviepy
imageio
imageio-ffmpeg
rembg
PyPDF2
